Fixed table and column names, data types
fixed some table and column names that didn't match schema, and some variables that had wrong data types

+++ Models/Review.py
@@ -6,7 +6,7 @@ import Spot
 class Review:
 
     reviewID: int
-    rating: float
+    rating: int
     noiseLevel: int
     content: str
     crowdedness: int
